refactor(step2): Position 클래스 추출
Car의 position 필드를 Wrapper 클래스로 추출, 불변객체화

+++ src/main/java/racingcar/view/GameResultView.java
@@ -33,7 +33,7 @@ public class GameResultView {
 
     private static String getPositionToString(Car car) {
         StringBuilder sb = new StringBuilder();
-        IntStream.range(0, car.getPosition()).forEach(i -> sb.append("-"));
+        IntStream.range(0, car.getPosition().toInt()).forEach(i -> sb.append("-"));
         return sb.toString();
     }
